requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
jsonschema>=4.20.0
python-pptx>=0.6.0
psycopg2-binary>=2.9.0
python-ulid>=2.0.0
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=21.0.0
requests>=2.28.0
PyYAML>=6.0.0
GitPython>=3.1.0
